split map routes into focused modules and converted to an aggregator
